Jar file to add word lists to database.
this is an executable jar file that has the sql libraries packaged into it. It looks for a database at a place you have specified and adds words into the tables words1 up to words5. If the tables do not exist, it will create them. Then it looks for 5 text files with the same names in a folder that you specify and adds all of the words into the tables. Words in the text files must be one word per line, separated with new lines.