Difficulty selection is now functional
The user can pause the game and select from one of four difficulties on
the pause menu. I had to move the pause function and split it up a bit
so there is now a function that pauses and a function that unpauses. The
asteroids speed up or slow down depending on difficulty